test: add perf suite — tackBody benches + budgets + bignum side-by-side
Performance testing for the migrated tick body and bignum library, with
a detailed report formatter.

## Tick perf (D/E/F fixtures)

Extracted the D/E/F fixture definitions, helpers, and runBodyTicks
driver from tickHarness.parity.test.ts into a shared module so parity,
bench, and budget tests all exercise the exact same shapes (-586 lines
in the parity test).

* test/fixtures/tackBodyFixtures.ts — shared fixture module
* test/perf/tickHarness.bench.ts — vitest bench(), one per fixture,
  auto-iterates and reports hz / mean / median / min / max / p99 /
  std-dev / rme / sample count.
* test/perf/tickHarness.budget.test.ts — median-of-7 budget
  assertions (3 warmup + 7 timed). Budgets D 15ms, E 20ms, F 15ms;
  ~8x headroom over local baselines (D 1.77ms, E 2.37ms, F 0.96ms)
  to absorb CI jitter without rewarding cherry-picked best cases.

## Bignum bench

Side-by-side break_infinity.js vs break_eternity.js (the latter
installed as a new dep) across construct / add / mul / div / pow /
log10 / exp / eq / gt / toString at four magnitude tiers (small 1e3 /
medium 1e100 / large 1e300 / very_large 1e1e6), plus a
break_eternity-only tetration tier (10^^100) showing what eternity
costs at its native range.

Observed: break_infinity is ~2-10x faster on arithmetic across all
tiers; break_eternity wins on `exp` and slightly on `toString`. Both
handle the test magnitudes without precision loss.

* test/perf/bignum.bench.ts
* packages/logic dep: break_eternity.js ^2.1.3

## Report tooling

`npm run bench:report` runs `vitest bench --outputJson` then pipes
the JSON through scripts/bench-report.mjs, which formats per-benchmark
console tables and writes BENCH_REPORT.md with full stats: ops/sec,
mean, median, min, max, p75, p99, p99.5, p99.9, std-dev, rme, sample
count. Plus per-group speed ranking, file-level aggregates, top-5
noisiest and slowest cross-file summaries. Includes Node + OS labels
and ISO timestamp for reproducibility tracking.

* scripts/bench-report.mjs — formatter
* package.json scripts: `bench` (vitest default), `bench:report`
* vitest.config.ts: `benchmark.include: ['test/**/*.bench.ts']`
* .gitignore: bench-results.json + BENCH_REPORT.md (regeneratable)

## Verification

* `npm run check:tsc` — clean across logic + web_ui + desktop_ui.
* `npm --workspace @synergism/logic test` — 99 files, 32,827 tests
  passing (was 32,824; +3 budget assertions).
* `npm --workspace @synergism/logic run bench:report` — both bench
  files run end-to-end, report generated, console + markdown outputs
  validated.
