+++ .gitignore
@@ -36,3 +36,7 @@ dist/dist/
 
 # Claude Code local-only settings (per-user permission allowlist, etc.)
 .claude/settings.local.json
+
+# Vitest bench artifacts (regenerated by `npm run bench:report`)
+packages/logic/bench-results.json
+packages/logic/BENCH_REPORT.md
