+++ packages/logic/package.json
@@ -12,9 +12,12 @@
   "scripts": {
     "build": "tsc --build",
     "check:tsc": "tsc --noEmit",
-    "test": "vitest run"
+    "test": "vitest run",
+    "bench": "vitest bench --run",
+    "bench:report": "vitest bench --run --outputJson=bench-results.json && node scripts/bench-report.mjs"
   },
   "dependencies": {
+    "break_eternity.js": "^2.1.3",
     "break_infinity.js": "^2.0.0",
     "fast-mersenne-twister": "^1.0.3",
     "lz-string": "^1.4.4",
